# inspection/upper_computer/frontend/openapi/inspection_gateway_openapi.json
{
  "components": {
...
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Inspection HMI Gateway",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/station/snapshot": {
      "get": {
        "description": "Return the current projected station snapshot.",
        "operationId": "getStationSnapshot",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getstationsnapshot",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get the station snapshot",
        "tags": [
          "station"
        ]
      }
    },
    "/api/v1/station/stats": {
      "get": {
        "description": "Return the current projected station counter payload.",
        "operationId": "getStationStats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Getstationstats",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get station counters",
        "tags": [
          "station"
        ]
      }
    },
...
  }
}